+++ logs/error/.5f1151ced196762ae8720e2c85f1519895bd259e-audit.json
@@ -1,0 +1,15 @@
+{
+    "keep": {
+        "days": false,
+        "amount": 30
+    },
+    "auditLog": "logs\\error\\.5f1151ced196762ae8720e2c85f1519895bd259e-audit.json",
+    "files": [
+        {
+            "date": 1711224272485,
+            "name": "logs\\error\\2024-03-24.error.log",
+            "hash": "ef88e0075f66bf729f80cb99414cece5d72316e1cefc3ceefa676870222242c8"
+        }
+    ],
+    "hashType": "sha256"
+}
