+++ package.json
@@ -19,7 +19,8 @@
         "nodemailer": "^6.9.13",
         "nodemailer-smtp-transport": "^2.7.4",
         "prisma": "^5.11.0",
-        "winston": "^3.12.0"
+        "winston": "^3.12.0",
+        "winston-daily-rotate-file": "^5.0.0"
     },
     "devDependencies": {
         "nodemon": "^3.1.0"
